% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_ids.R
\docType{data}
\name{character_ids}
\alias{character_ids}
\title{Characters associated with each id.}
\format{
A data frame with 731 rows and 2 variables:
\describe{
  \item{id}{unique id for ech character}
  \item{character}{name of the superhero}
  ...
}
}
\usage{
character_ids
}
\description{
A dataset containing the ids associated with each superhero
}
\keyword{datasets}
