% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bio.R
\name{get_bio}
\alias{get_bio}
\title{Get biographic information about a superhero.}
\usage{
get_bio(access_token = Sys.getenv("SUPERHERO_TOKEN"), character_id)
}
\arguments{
\item{access_token}{Required. Unique token obtained from the SuperHero API site. Token should be set to the environment and will be retrieved
that way.}

\item{character_id}{Required. Required. The \href{https://superheroapi.com/ids.html}{Character ID} or name of the Superhero}
}
\value{
Returns a dataframe with the work/occupation  of the specific superhero.
}
\description{
Get biographic information about a superhero.
}
