% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_powerstats.R
\name{get_powerstats}
\alias{get_powerstats}
\title{Retrieve powerstats for a superhero}
\usage{
get_powerstats(access_token = Sys.getenv("SUPERHERO_TOKEN"),
  character_id)
}
\arguments{
\item{access_token}{Unique token obtained from the SuperHero API site. Token should be set to the environment and will be retrieved
that way.}

\item{character_id}{Id number of name of superhero character.}
}
\value{
tibble
}
\description{
Retrieve powerstats for a superhero
}
\examples{
\dontrun{
 get_powerstats(character_id = 69)
}

}
