% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_work.R
\name{get_work}
\alias{get_work}
\title{Get work/occupation for superhero}
\usage{
get_work(access_token = Sys.getenv("SUPERHERO_TOKEN"), character_id)
}
\arguments{
\item{access_token}{Required. Unique token obtained from the SuperHero API site. Token should be set to the environment and will be retrieved
that way.}

\item{character_id}{Required. The \href{https://superheroapi.com/ids.html}{Character ID} or name of the Superhero}
}
\value{
Returns a dataframe with the work/occupation  of the specific superhero.
}
\description{
Get work/occupation for superhero
}
