% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superhero_complete.R
\docType{data}
\name{superhero_complete}
\alias{superhero_complete}
\title{Superhero metadata}
\format{
A data frame with 3662 rows and 26 variables:
\describe{
  \item{response}{unique id for ech character}
  \item{id}{name of the superhero}
  \item{name}superhero name{}
  \item{gender}{gender of superhero}
  \item{race}{race of superhero}
  \item{height}{height in feet and cm}
  \item{weight}{weight in kg and lb}
  \item{eye_color}{superhero eye color}
  \item{hair_color}{superhero hair color}
  \item{full_name}{full name of superhero}
  \item{alter_egos}{superhero alter egos}
  \item{aliases}{superhero aliases}
  \item{place_of_birth}{birth place of superhero}
  \item{first_appearance}{first appearance of superhero}
  \item{publisher}{publisher of comic }
  \item{aligment}{superheor alignment (bad, good, neutral, unkown)}
  \item{group_affilation}{superhero affliation}
  \item{relatives}{relatives to superhero}
  \item{intelligence}{superhero intelligence, 1-100}
  \item{strength}{superhero strength, 1-100}
  \item{speed}{superhero speed, 1-100}
  \item{durability}{superhero durability, 1-100}
  \item{power}{superhero power, 1-100}
  \item{combat}{superhero combat, 1-100}
  \item{occupation}{superhero occupatoin}
  \item{base}{superhero base(s)}
  ...
}
}
\usage{
superhero_complete
}
\description{
A dataset containing information about every superhero.
}
\keyword{datasets}
