% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_powerstats.R
\name{get_powerstats}
\alias{get_powerstats}
\title{Call that gets all the powerstats for the superhero.}
\usage{
get_powerstats(access_token = Sys.getenv("SUPERHERO_TOKEN"), character_id)
}
\arguments{
\item{access_token}{Required. Unique token obtained from the SuperHero API site. Token should be set to the environment and will be retrieved
that way.}

\item{character_id}{Required. The \href{https://superheroapi.com/ids.html}{Character ID} or name of the Superhero}
}
\value{
Returns a dataframe with the powerstats of the specific superhero.
}
\description{
Call that gets all the powerstats for the superhero.
}
\examples{
\dontrun{
 get_powerstats(character_id = 69)
}

}
